requirements.txt

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.3.2
python-dotenv==1.0.1
requests==2.32.3
shap==0.45.1
